% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derange.R
\name{derange}
\alias{derange}
\title{A shuffling Function so that all numbers from 1 through n are shuffled}
\usage{
derange(n)
}
\arguments{
\item{n}{A whole number greater than 0}
}
\description{
This function returns a shuffling of n numbers
in a random order so that the number does not
match the placement position
}
\examples{
derange()
}
\keyword{shuffling}
