% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nicePlot.R
\name{nicePlot}
\alias{nicePlot}
\title{A function that plots data frame into a line graph}
\usage{
nicePlot(df, color)
}
\arguments{
\item{c(df, }{color)}
}
\description{
This function returns plot of data frame with
specific color scheme of red, blue, or green.
}
\examples{
nicePlot()
}
\keyword{plot}
